# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERB System API Documentation",
        "description": "Tài liệu đặc tả và kiểm thử RESTful API - Hệ thống quản trị ERB",
        "contact": {
            "email": "admin@erb.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/reports/inventory-xnt": {
            "get": {
                "tags": [
                    "Report (Báo cáo XNT)"
                ],
                "summary": "Báo cáo Xuất - Nhập - Tồn theo kỳ",
                "description": "Tồn đầu kỳ + Nhập - Xuất (± điều chuyển) = Tồn cuối kỳ. scope=warehouse | system.",
                "operationId": "81e82e7e3542bd57c749194d74586881",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "scope",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "warehouse",
                                "system"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Thiếu/khoảng ngày sai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT Access Token (dạng: Bearer <token>) để thực hiện các yêu cầu cần xác thực.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}